refactor: split analytics DTOs into separate files
